refactor(ui-library): rename CosPercentagePieChart.tsx folder
