+++ packages/cube-frontend-ui-library/src/components/CosPercentagePieChart/CosPercentagePieChart.tsx
