+++ packages/cube-frontend-ui-library/src/components/CosPercentagePieChart/PercentagePie.tsx
